refactor: move commands to individual directories
- Moved cli/cli.py to cli.py and set as main entrypoint in
pyproject.toml.
- Moved each commands CLI wrapper from cli/<command>.py to
commands/<command>/cli.py and the logic script from
commands/<command.py> to commands/<command>/<command>.py.
- Updated all import statements.
- Removed main.py (as replaced with cli.py).
