+++ src/evaluator/__main__.py
@@ -4,5 +4,5 @@ __main__.py
 Adds support for running via: python -m ...
 '''
 
-from evaluator.main import evaluator
+from evaluator.cli import evaluator
 evaluator()
