+++ src/evaluator/commands/analyse/analyse.py
