+++ src/evaluator/commands/license/license.py
